+++ CHANGES.md
@@ -1,5 +1,16 @@
 # **ANGoLS** Changes
 
+## 0.6.0 (alpha 1) - 24th February 2025
+
+* refactored all slice functions into **slices** package;
+* refactored all slice functions into **strings** package;
+* added `strings.IndexAfter()`;
+* added `strings.IndexAnyAfter()`;
+* added `strings.IndexByteAfter()`;
+* added `strings.IndexFuncAfter()`;
+* added `strings.IndexNotAnyAfter()`;
+
+
 ## 0.5.0 - 30th March 2019
 
 * + added ``StringChomp()`` and ``StringChompAll()`` functions